Rename package utility to db
