+++ src/checkpoint/andela/db/DBWriter.java
@@ -1,0 +1,4 @@
+package checkpoint.andela.db;
+
+public class DBWriter {
+}
