+++ src/checkpoint/andela/db/DbRecord.java
@@ -1,4 +1,4 @@
-package checkpoint.andela.utility;
+package checkpoint.andela.db;
 
 import java.util.ArrayList;
 import java.util.HashMap;
